requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiofiles>=23.0.0
APScheduler>=3.10.4
flask>=3.0.0
python-dateutil>=2.8.2
